add inbound port definition
