+++ AnalyticsServiceChangeFeedInboundProcessorAdapter/function.json
