+++ AnalyticsServiceChangeFeedInboundProcessorAdapter/index.ts
@@ -1,13 +1,14 @@
 import { Context } from "@azure/functions";
+import { QueueClient } from "@azure/storage-queue";
 import { RetrievedService } from "@pagopa/io-functions-commons/dist/src/models/service";
 import * as KP from "../utils/kafka/KafkaProducerCompact";
 import { ValidableKafkaProducerConfig } from "../utils/kafka/KafkaTypes";
 import { getConfigOrThrow } from "../utils/config";
 import * as KA from "../outbound/adapter/kafka-outbound-publisher";
-import * as EA from "../outbound/adapter/throw-outbound-publisher";
+import * as QA from "../outbound/adapter/queue-outbound-publisher";
 import * as TA from "../outbound/adapter/tracker-outbound-publisher";
+import { getAnalyticsProcessorForService } from "../businesslogic/analytics-services";
 import { OutboundPublisher } from "../outbound/port/outbound-publisher";
-import { processService } from "../businesslogic/analytics-services";
 import { avroServiceFormatter } from "../utils/formatter/servicesAvroFormatter";
 
 const config = getConfigOrThrow();
@@ -23,18 +24,25 @@ const retrievedServiceOnKafkaAdapter: OutboundPublisher<RetrievedService> = KA.c
   )
 );
 
-const throwAdapter: OutboundPublisher<RetrievedService> = EA.create();
+const retrievedServiceOnQueueAdapter: OutboundPublisher<RetrievedService> = QA.create(
+  new QueueClient(
+    config.INTERNAL_STORAGE_CONNECTION_STRING,
+    config.SERVICES_FAILURE_QUEUE_NAME
+  )
+);
 
 const telemetryAdapter = TA.create(
   TA.initTelemetryClient(config.APPINSIGHTS_INSTRUMENTATIONKEY)
 );
 
-const run = (_context: Context, document: unknown): Promise<void> =>
-  processService(
+const run = (
+  _context: Context,
+  documents: ReadonlyArray<unknown>
+): Promise<void> =>
+  getAnalyticsProcessorForService(
     telemetryAdapter,
     retrievedServiceOnKafkaAdapter,
-    throwAdapter,
-    [document]
-  )();
+    retrievedServiceOnQueueAdapter
+  ).process(documents)();
 
 export default run;
