+++ AnalyticsServiceStorageQueueInboundProcessorAdapter/function.json
